+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added a `last_used_at` field to API token.
 - Added an upload progress bar to the CLI.
+- Added an upload progress bar to the web interface.
 
 ### Changed